backend/requirements/common.txt
black
django~=3.0.0
djangorestframework
django-cors-headers
drf-jwt
httpie
pillow
django-pydenticon